gateway-auth/requirements.txt
# FastAPI JWT Authentication Dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
redis==5.1.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.0
slowapi==0.1.9
cryptography==43.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
fakeredis==2.26.1
freezegun==1.5.1